[#28] refactor : management 입력 custom hook으로 정리
