+++ catch-study/src/pages/ManagementInfoPage/ManagementInfo.tsx
@@ -1,123 +1,26 @@
-import { ChangeEvent, useState } from 'react';
 import AddressForm from '../../components/AddressForm/AddreesFrom';
 import FeeForm from '../../components/FeeForm/FeeForm';
 import RoomForm from '../../components/RoomForm/RoomForm';
 import Topbar from '../../components/Topbar/Topbar';
-import {
-  FormDataTypes,
-  RoomInfoTypes,
-  UsageFeeTypes,
-} from '../../types/manage';
+import useManagementInfo from '../../hooks/useManagementInfo';
 
 const ManagementInfo: React.FC = () => {
-  const [roomInfos, setRoomInfos] = useState<RoomInfoTypes[]>([]);
-  const [cancelTime, setCancelTime] = useState(0);
-  const [usageFees, setUsageFees] = useState<UsageFeeTypes[]>([]);
-
-  const [formData, setFormData] = useState<FormDataTypes>({
-    cafeName: '',
-    address: {
-      city: '',
-      country: '',
-      town: '',
-      etc: '',
-    },
-    openingHours: '',
-    closedHours: '',
-    closedDay: '',
-    seats: 0,
-    roomInfo: [
-      {
-        cancelAvailableTime: cancelTime,
-        rooms: roomInfos,
-      },
-    ],
-    usageFee: usageFees,
-    titleImage: '',
-    multipleImages: [],
-    seatChartImage: '',
-    cafePhone: '',
-  });
-
-  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-
-    setFormData(prevFormData => ({
-      ...prevFormData,
-      [name]: value,
-    }));
-  };
-
-  const handleNestedInputChange = (
-    e: ChangeEvent<HTMLInputElement>,
-    field: string,
-    setData: React.Dispatch<React.SetStateAction<FormDataTypes>>,
-  ) => {
-    const { name, value } = e.target;
-
-    setData(prevFormData => ({
-      ...prevFormData,
-      [field]: {
-        ...prevFormData.address,
-        [name]: value,
-      },
-    }));
-  };
-
-  const handleRoomChange = (
-    index: number,
-    field: keyof RoomInfoTypes,
-    value: string,
-  ) => {
-    const numberValue = value === '' ? 0 : parseInt(value, 10);
-
-    setRoomInfos(prev =>
-      prev.map((room, i) =>
-        i === index ? { ...room, [field]: numberValue } : room,
-      ),
-    );
-  };
-
-  const handleRoomNameChange = (
-    e: ChangeEvent<HTMLInputElement>,
-    index: number,
-  ) => {
-    const { name, value } = e.target;
-
-    setRoomInfos(prev =>
-      prev.map((room, i) => (i === index ? { ...room, [name]: value } : room)),
-    );
-  };
-
-  const handleAddRoom = () => {
-    setRoomInfos(prev => [...prev, { name: '', capacity: 0, counts: 0 }]);
-  };
-
-  const handleRemoveRoom = (index: number) => {
-    setRoomInfos(prev => prev.filter((_, i) => i !== index));
-  };
-
-  const handleAddFee = () => {
-    setUsageFees(prev => [...prev, { hours: 0, price: 0 }]);
-  };
-
-  const handleRemoveFee = (index: number) => {
-    setUsageFees(prev => prev.filter((_, i) => i !== index));
-  };
-
-  const handleFeeChange = (
-    index: number,
-    field: keyof UsageFeeTypes,
-    value: string,
-  ) => {
-    const numberValue = value === '' ? 0 : parseInt(value, 10);
-
-    setUsageFees(prev =>
-      prev.map((room, i) =>
-        i === index ? { ...room, [field]: numberValue } : room,
-      ),
-    );
-  };
+  const {
+    roomInfos,
+    usageFees,
+    formData,
+    setFormData,
+    setCancelTime,
+    handleInputChange,
+    handleNestedInputChange,
+    handleRoomChange,
+    handleRoomNameChange,
+    handleAddRoom,
+    handleRemoveRoom,
+    handleAddFee,
+    handleRemoveFee,
+    handleFeeChange,
+  } = useManagementInfo();
 
   return (
     <div className='w-screen h-full'>
